Changed re-rendering method after delete
