+++ src/utils/utils.ts
@@ -1,0 +1,9 @@
+import { getTasks } from '../modules/getTasks';
+import bind from '../binds';
+
+export const renderAndBindTaskList = () => {
+    getTasks().then((taskList) => {
+        bind.editTaskButton(taskList);
+        bind.removeTaskButton(taskList);
+    });
+}
